fix: compile error

+++ a_sync/primitives/locks/counter.pxd
@@ -6,7 +6,7 @@ cdef class CounterLock(_DebugDaemonMixin):
     cdef char* __name
     cdef long long _value
     cdef dict[long long, Event] _events
-    cpdef bint is_ready(self, int v)
-    cdef bint c_is_ready(self, int v)
-    cpdef void set(self, int value)
-    cdef void c_set(self, int value)
+    cpdef bint is_ready(self, long long v)
+    cdef bint c_is_ready(self, long long v)
+    cpdef void set(self, long long value)
+    cdef void c_set(self, long long value)
